+++ crypto/multisig/src/constants.rs
@@ -1,0 +1,4 @@
+// Copyright (c) 2018-2022 The MobileCoin Foundation
+
+/// The maximum number of signatures that can be included in a multi-signature.
+pub const MAX_SIGNATURES: usize = 10;
